new AbstractBaseTheme and BaseSeries; introducing base font concept for Theme; new method for setting sum size in pie charts
