+++ xchart/src/main/java/org/knowm/xchart/style/MatlabTheme.java
@@ -22,7 +22,6 @@ import java.awt.Font;
 import java.awt.Stroke;
 
 import org.knowm.xchart.style.PieStyler.AnnotationType;
-import org.knowm.xchart.style.Styler.LegendPosition;
 import org.knowm.xchart.style.colors.ChartColor;
 import org.knowm.xchart.style.colors.MatlabSeriesColors;
 import org.knowm.xchart.style.lines.MatlabSeriesLines;
@@ -32,27 +31,11 @@ import org.knowm.xchart.style.markers.MatlabSeriesMarkers;
 /**
  * @author timmolter
  */
-public class MatlabTheme implements Theme {
+public class MatlabTheme extends AbstractBaseTheme {
 
   // Chart Style ///////////////////////////////
 
-  @Override
-  public Color getChartBackgroundColor() {
-
-    return ChartColor.getAWTColor(ChartColor.WHITE);
-  }
-
-  @Override
-  public Color getChartFontColor() {
-
-    return ChartColor.getAWTColor(ChartColor.BLACK);
-  }
-
-  @Override
-  public int getChartPadding() {
-
-    return 10;
-  }
+  // SeriesMarkers, SeriesLines, SeriesColors ///////////////////////////////
 
   @Override
   public Marker[] getSeriesMarkers() {
@@ -75,60 +58,12 @@ public class MatlabTheme implements Theme {
   // Chart Title ///////////////////////////////
 
   @Override
-  public Font getChartTitleFont() {
-
-    return new Font(Font.SANS_SERIF, Font.BOLD, 14);
-  }
-
-  @Override
-  public boolean isChartTitleVisible() {
-
-    return true;
-  }
-
-  @Override
   public boolean isChartTitleBoxVisible() {
 
     return false;
   }
 
-  @Override
-  public Color getChartTitleBoxBackgroundColor() {
-
-    return ChartColor.getAWTColor(ChartColor.WHITE);
-  }
-
-  @Override
-  public Color getChartTitleBoxBorderColor() {
-
-    return ChartColor.getAWTColor(ChartColor.WHITE);
-  }
-
-  @Override
-  public int getChartTitlePadding() {
-
-    return 5;
-  }
-
   // Chart Legend ///////////////////////////////
-
-  @Override
-  public Font getLegendFont() {
-
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 11);
-  }
-
-  @Override
-  public boolean isLegendVisible() {
-
-    return true;
-  }
-
-  @Override
-  public Color getLegendBackgroundColor() {
-
-    return ChartColor.getAWTColor(ChartColor.WHITE);
-  }
 
   @Override
   public Color getLegendBorderColor() {
@@ -136,78 +71,18 @@ public class MatlabTheme implements Theme {
     return ChartColor.getAWTColor(ChartColor.BLACK);
   }
 
-  @Override
-  public int getLegendPadding() {
-
-    return 10;
-  }
-
-  @Override
-  public int getLegendSeriesLineLength() {
-
-    return 24;
-  }
-
-  @Override
-  public LegendPosition getLegendPosition() {
-
-    return LegendPosition.OutsideE;
-  }
-
   // Chart Axes ///////////////////////////////
-
-  @Override
-  public boolean isXAxisTitleVisible() {
-
-    return true;
-  }
-
-  @Override
-  public boolean isYAxisTitleVisible() {
-
-    return true;
-  }
 
   @Override
   public Font getAxisTitleFont() {
 
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 12);
-  }
-
-  @Override
-  public boolean isXAxisTicksVisible() {
-
-    return true;
-  }
-
-  @Override
-  public boolean isYAxisTicksVisible() {
-
-    return true;
-  }
-
-  @Override
-  public Font getAxisTickLabelsFont() {
-
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 12);
+    return getBaseFont().deriveFont(12f);
   }
 
   @Override
   public int getAxisTickMarkLength() {
 
     return 5;
-  }
-
-  @Override
-  public int getAxisTickPadding() {
-
-    return 4;
-  }
-
-  @Override
-  public int getPlotMargin() {
-
-    return 3;
   }
 
   @Override
@@ -223,12 +98,6 @@ public class MatlabTheme implements Theme {
   }
 
   @Override
-  public Color getAxisTickLabelsColor() {
-
-    return ChartColor.getAWTColor(ChartColor.BLACK);
-  }
-
-  @Override
   public boolean isAxisTicksLineVisible() {
 
     return false;
@@ -240,66 +109,12 @@ public class MatlabTheme implements Theme {
     return false;
   }
 
-  @Override
-  public int getAxisTitlePadding() {
-
-    return 10;
-  }
-
-  @Override
-  public int getXAxisTickMarkSpacingHint() {
-
-    return 74;
-  }
-
-  @Override
-  public int getYAxisTickMarkSpacingHint() {
-
-    return 44;
-  }
-
   // Chart Plot Area ///////////////////////////////
-
-  @Override
-  public boolean isPlotGridLinesVisible() {
-
-    return true;
-  }
-
-  @Override
-  public boolean isPlotGridVerticalLinesVisible() {
-
-    return true;
-  }
-
-  @Override
-  public boolean isPlotGridHorizontalLinesVisible() {
-
-    return true;
-  }
-
-  @Override
-  public Color getPlotBackgroundColor() {
-
-    return ChartColor.getAWTColor(ChartColor.WHITE);
-  }
 
   @Override
   public Color getPlotBorderColor() {
 
     return ChartColor.getAWTColor(ChartColor.BLACK);
-  }
-
-  @Override
-  public boolean isPlotBorderVisible() {
-
-    return true;
-  }
-
-  @Override
-  public boolean isPlotTicksMarksVisible() {
-
-    return true;
   }
 
   @Override
@@ -315,50 +130,14 @@ public class MatlabTheme implements Theme {
   }
 
   @Override
-  public double getPlotContentSize() {
+  public int getPlotMargin() {
 
-    return .92;
+    return 3;
   }
 
   // Category Charts ///////////////////////////////
 
-  @Override
-  public double getAvailableSpaceFill() {
-
-    return 0.9;
-  }
-
-  @Override
-  public boolean isOverlapped() {
-
-    return false;
-  }
-
   // Pie Charts ///////////////////////////////
-
-  @Override
-  public boolean isCircular() {
-
-    return true;
-  }
-
-  @Override
-  public double getStartAngleInDegrees() {
-
-    return 0;
-  }
-
-  @Override
-  public Font getPieFont() {
-
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 15);
-  }
-
-  @Override
-  public double getAnnotationDistance() {
-
-    return .67;
-  }
 
   @Override
   public AnnotationType getAnnotationType() {
@@ -366,37 +145,7 @@ public class MatlabTheme implements Theme {
     return AnnotationType.Label;
   }
 
-  @Override
-  public boolean isDrawAllAnnotations() {
-
-    return false;
-  }
-
-  @Override
-  public double getDonutThickness() {
-
-    return .33;
-  }
-
-  @Override
-  public boolean isSumVisible() {
-
-  	return false;
-  }
-
-  @Override
-  public Font getSumFont() {
-
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 12);
-  }
-
   // Line, Scatter, Area Charts ///////////////////////////////
-
-  @Override
-  public int getMarkerSize() {
-
-    return 8;
-  }
 
   @Override
   public boolean showMarkers() {
@@ -406,23 +155,6 @@ public class MatlabTheme implements Theme {
 
   // Error Bars ///////////////////////////////
 
-  @Override
-  public Color getErrorBarsColor() {
-
-    return ChartColor.getAWTColor(ChartColor.BLACK);
-  }
-
-  @Override
-  public boolean isErrorBarsColorSeriesColor() {
-
-    return false;
-  }
-
   // Annotations ///////////////////////////////
 
-  @Override
-  public Font getAnnotationFont() {
-
-    return new Font(Font.SANS_SERIF, Font.PLAIN, 12);
-  }
 }
