+++ xchart/src/main/java/org/knowm/xchart/style/Theme.java
@@ -33,6 +33,13 @@ public interface Theme extends SeriesMarkers, SeriesLines, SeriesColors {
 
   // Chart Style ///////////////////////////////
 
+  /**
+   * Base font, from which all other fonts are derived, if not set directly.
+   *
+   * @return base font
+   */
+  Font getBaseFont();
+
   Color getChartBackgroundColor();
 
   Color getChartFontColor();
