+++ xchart/src/main/java/org/knowm/xchart/style/lines/BaseSeriesLines.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright 2015-2017 Knowm Inc. (http://knowm.org) and contributors.
+ * Copyright 2011-2015 Xeiam LLC (http://xeiam.com) and contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.knowm.xchart.style.lines;
+
+import java.awt.BasicStroke;
+
+/**
+ * @author timmolter
+ * @author ekleinod
+ */
+public class BaseSeriesLines implements SeriesLines {
+
+  private final BasicStroke[] seriesLines;
+
+  /**
+   * Constructor
+   */
+  public BaseSeriesLines() {
+
+    seriesLines = new BasicStroke[]{SOLID, DOT_DOT, DASH_DASH, DASH_DOT};
+  }
+
+  @Override
+  public BasicStroke[] getSeriesLines() {
+
+    return seriesLines;
+  }
+}
