+++ xchart/src/main/java/org/knowm/xchart/style/markers/BaseSeriesMarkers.java
@@ -1,0 +1,40 @@
+/**
+ * Copyright 2015-2017 Knowm Inc. (http://knowm.org) and contributors.
+ * Copyright 2011-2015 Xeiam LLC (http://xeiam.com) and contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.knowm.xchart.style.markers;
+
+/**
+ * @author timmolter
+ * @author ekleinod
+ */
+public class BaseSeriesMarkers implements SeriesMarkers {
+
+  private final Marker[] seriesMarkers;
+
+  /**
+   * Constructor
+   */
+  public BaseSeriesMarkers() {
+
+    seriesMarkers = new Marker[]{CIRCLE, SQUARE, DIAMOND, TRIANGLE_UP, TRIANGLE_DOWN};
+  }
+
+  @Override
+  public Marker[] getSeriesMarkers() {
+
+    return seriesMarkers;
+  }
+}
